# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * VARNOX-MD-V2 WhatsApp Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatusState = typeof BotStatusState[keyof typeof BotStatusState];


export const BotStatusState = {
  connected: 'connected',
  disconnected: 'disconnected',
  connecting: 'connecting',
  qr_ready: 'qr_ready',
  pair_code_ready: 'pair_code_ready',
} as const;

export interface BotStatus {
  state: BotStatusState;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  uptime?: number | null;
  /** @nullable */
  connectedAt?: string | null;
}

export interface PairCodeInput {
  /** Phone number with country code (e.g. +33612345678) */
  phone: string;
}

export interface PairCodeResult {
  /** 8-character pair code (format XXXX-XXXX) */
  code: string;
  /** Seconds until the code expires */
  expiresIn: number;
}

export interface BotInfo {
  /** @nullable */
  jid?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  platform?: string | null;
  version: string;
  uptime: number;
}

export interface DashboardStats {
  totalGroups: number;
  totalUsers: number;
  commandsUsed: number;
  messagesHandled: number;
  uptime: number;
  memoryUsage: number;
  cpuUsage: number;
  premiumUsers: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  command: 'command',
  join: 'join',
  leave: 'leave',
  error: 'error',
  message: 'message',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  timestamp: string;
  /** @nullable */
  user?: string | null;
  /** @nullable */
  group?: string | null;
}

export interface Command {
  name: string;
  description: string;
  category: string;
  usage: string;
  aliases: string[];
  isOwnerOnly: boolean;
  isGroupOnly: boolean;
  isPremium: boolean;
  usageCount: number;
}

export interface Group {
  jid: string;
  name: string;
  memberCount: number;
  isAdmin: boolean;
  antilink: boolean;
  welcome: boolean;
  muted: boolean;
}

export type BotUserRole = typeof BotUserRole[keyof typeof BotUserRole];


export const BotUserRole = {
  owner: 'owner',
  superadmin: 'superadmin',
  admin: 'admin',
  premium: 'premium',
  user: 'user',
} as const;

export interface BotUser {
  jid: string;
  /** @nullable */
  name?: string | null;
  role: BotUserRole;
  commandCount: number;
  warns: number;
  isPremium: boolean;
  joinedAt: string;
  level: number;
  exp: number;
  coins: number;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  superadmin: 'superadmin',
  admin: 'admin',
  premium: 'premium',
  user: 'user',
} as const;

export interface UserUpdate {
  role?: UserUpdateRole;
  isPremium?: boolean;
  warns?: number;
  coins?: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetCommandsParams = {
/**
 * Filter by category
 */
category?: string;
};

export type GetUsersParams = {
/**
 * Filter by role (owner, admin, premium, user)
 */
role?: string;
};