# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VARNOX-MD-V2 WhatsApp Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/pair-code:
    post:
      operationId: requestPairCode
      tags: [bot]
      summary: Request a WhatsApp pair code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairCodeInput"
      responses:
        "200":
          description: Pair code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairCodeResult"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Bot already connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all bot users
      parameters:
        - in: query
          name: role
          schema:
            type: string
          description: Filter by role (owner, admin, premium, user)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotUser"
  /users/{jid}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user
      parameters:
        - in: path
          name: jid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotUser"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role or premium status
      parameters:
        - in: path
          name: jid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotUser"
components:
  schemas:
...
    PairCodeInput:
      type: object
      properties:
        phone:
          type: string
          description: Phone number with country code (e.g. +33612345678)
      required:
        - phone
    PairCodeResult:
      type: object
      properties:
        code:
          type: string
          description: 8-character pair code (format XXXX-XXXX)
        expiresIn:
          type: number
          description: Seconds until the code expires
      required:
        - code
        - expiresIn
...
    BotUser:
      type: object
      properties:
        jid:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, superadmin, admin, premium, user]
        commandCount:
          type: number
        warns:
          type: number
        isPremium:
          type: boolean
        joinedAt:
          type: string
        level:
          type: number
        exp:
          type: number
        coins:
          type: number
      required:
        - jid
        - role
        - commandCount
        - warns
        - isPremium
        - joinedAt
        - level
        - exp
        - coins
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [superadmin, admin, premium, user]
        isPremium:
          type: boolean
        warns:
          type: number
        coins:
          type: number
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error